Add PROXY protocol unique ID TLV helpers
