+++ actix-proxy-protocol/src/tlv.rs
@@ -170,48 +170,8 @@ impl Tlv for Noop {
     }
 }
 
-/// The value of the type PP2_TYPE_UNIQUE_ID is an opaque byte sequence of up to
-/// 128 bytes generated by the upstream proxy that uniquely identifies the
-/// connection.
-///
-/// The unique ID can be used to easily correlate connections across multiple
-/// layers of proxies, without needing to look up IP addresses and port numbers.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct UniqueId {
-    value: Vec<u8>,
-}
-
-impl UniqueId {
-    ///
-    ///
-    /// # Panics
-    /// Panics if `value` is 0 bytes or larger than 128 bytes.
-    pub fn new(id: impl Into<Vec<u8>>) -> Self {
-        let value = id.into();
-
-        assert!(!value.is_empty(), "UniqueId TLV `value` cannot be empty");
-        assert!(
-            value.len() <= 128,
-            "UniqueId TLV `value` cannot be larger than 128 bytes"
-        );
-
-        Self { value }
-    }
-}
-
-impl Tlv for UniqueId {
-    const TYPE: u8 = PP2_TYPE_UNIQUE_ID;
-
-    fn try_from_value(value: &[u8]) -> Option<Self> {
-        Some(Self {
-            value: value.to_owned(),
-        })
-    }
-
-    fn value_bytes(&self) -> Cow<'_, [u8]> {
-        Cow::Borrowed(&self.value)
-    }
-}
+mod unique_id;
+pub use self::unique_id::UniqueId;
 
 bitflags::bitflags! {
     #[derive(Debug, Clone, PartialEq, Eq)]
