+++ actix-proxy-protocol/src/tlv/unique_id.rs
@@ -1,0 +1,87 @@
+use std::borrow::Cow;
+
+use super::{PP2_TYPE_UNIQUE_ID, Tlv};
+
+/// The value of the type PP2_TYPE_UNIQUE_ID is an opaque byte sequence of up to
+/// 128 bytes generated by the upstream proxy that uniquely identifies the
+/// connection.
+///
+/// The unique ID can be used to easily correlate connections across multiple
+/// layers of proxies, without needing to look up IP addresses and port numbers.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct UniqueId {
+    value: Vec<u8>,
+}
+
+impl UniqueId {
+    /// Constructs a new unique ID TLV.
+    ///
+    /// # Panics
+    /// Panics if `value` is 0 bytes or larger than 128 bytes.
+    pub fn new(id: impl Into<Vec<u8>>) -> Self {
+        let value = id.into();
+
+        assert!(!value.is_empty(), "UniqueId TLV `value` cannot be empty");
+        assert!(
+            value.len() <= 128,
+            "UniqueId TLV `value` cannot be larger than 128 bytes"
+        );
+
+        Self { value }
+    }
+
+    /// Returns the unique ID bytes.
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.value
+    }
+}
+
+impl Tlv for UniqueId {
+    const TYPE: u8 = PP2_TYPE_UNIQUE_ID;
+
+    fn try_from_value(value: &[u8]) -> Option<Self> {
+        Some(Self {
+            value: value.to_owned(),
+        })
+    }
+
+    fn value_bytes(&self) -> Cow<'_, [u8]> {
+        Cow::Borrowed(&self.value)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn unique_id_tlv_round_trip() {
+        let unique_id = UniqueId::new("abc123");
+
+        assert_eq!(unique_id.as_bytes(), b"abc123");
+        assert_eq!(unique_id.value_bytes(), b"abc123".as_slice());
+        assert_eq!(UniqueId::try_from_value(b"abc123"), Some(unique_id.clone()));
+        assert_eq!(UniqueId::try_from_parts(0x05, b"abc123"), Some(unique_id));
+        assert_eq!(UniqueId::try_from_parts(0x04, b"abc123"), None);
+    }
+
+    #[test]
+    fn unique_id_tlv_accepts_128_byte_constructor_value() {
+        let value = vec![0x42; 128];
+        let unique_id = UniqueId::new(value.clone());
+
+        assert_eq!(unique_id.as_bytes(), value.as_slice());
+    }
+
+    #[test]
+    #[should_panic = "UniqueId TLV `value` cannot be empty"]
+    fn unique_id_tlv_rejects_empty_constructor_value() {
+        UniqueId::new(Vec::new());
+    }
+
+    #[test]
+    #[should_panic = "UniqueId TLV `value` cannot be larger than 128 bytes"]
+    fn unique_id_tlv_rejects_oversized_constructor_value() {
+        UniqueId::new(vec![0x42; 129]);
+    }
+}
